requirements.txt
django
djangorestframework
django-rest-swagger
cryptography
python-dotenv
